Update README.md
This repository contains comprehensive notes and insights from my Qualys Policy Compliance Foundation certification course. It covers the essentials of IT policy compliance, including applicable regulations, best practices, and the importance of automation. The course aims to equip individuals with the knowledge to design, implement, and manage effective IT compliance programs, ensuring both internal and external standards are met. Dive in to explore the frameworks, regulations, and strategies for achieving robust IT policy compliance.